Update Action UI to use dynamic properties when updating the icon.
+ Left some comments attempting to use QKeySequenceEdit.

Note that using QKeySequenceEdit has limitation for this app use case,
because it treats key modifiers (i.e. Ctrl, Alt) not as a regular key,
so it have to come in combination with another key (i.e. Ctrl+F5).
This has limitation, since we can't fully map a key to another key, i.e.
making Ctrl work as Alt key.